Integrate OC-DECLARE with crate re-structuring
