+++ process_mining/examples/oc_declare.rs
@@ -1,8 +1,13 @@
 use std::{env::args, path::PathBuf};
 
 use process_mining::{
-    import_ocel_xml_file,
-    object_centric::oc_declare::{self, OCDeclareDiscoveryOptions},
+    core::{
+        event_data::object_centric::ocel_xml::import_ocel_xml_file,
+        process_models::object_centric::oc_declare,
+    },
+    discovery::object_centric::oc_declare::{
+        discover_behavior_constraints, OCDeclareDiscoveryOptions,
+    },
 };
 
 pub fn main() {
@@ -10,7 +15,7 @@ pub fn main() {
     if let Some(path) = path_opt.map(PathBuf::from) {
         let ocel = import_ocel_xml_file(&path);
         let processed_locel = oc_declare::preprocess_ocel(ocel);
-        let discovered_constraints = oc_declare::discover_behavior_constraints(
+        let discovered_constraints = discover_behavior_constraints(
             &processed_locel,
             OCDeclareDiscoveryOptions {
                 ..Default::default()
