+++ process_mining/src/conformance/mod.rs
@@ -4,3 +4,5 @@
 //! event data.
 pub mod case_centric;
 pub mod object_centric;
+pub use case_centric::*;
+pub use object_centric::*;
