+++ process_mining/src/conformance/object_centric/mod.rs
@@ -1,2 +1,3 @@
 //! Object-centric Conformance-checking Approaches
 pub mod object_centric_language_abstraction;
+pub mod oc_declare;
