+++ process_mining/src/core/process_models/mod.rs
@@ -1,3 +1,5 @@
 //! Base Structs and Utils for Process Models
 pub mod case_centric;
 pub mod object_centric;
+pub use case_centric::*;
+pub use object_centric::*;
