+++ process_mining/src/core/process_models/object_centric/oc_declare/preprocess.rs
@@ -1,9 +1,9 @@
 use itertools::Itertools;
 
-use crate::{
-    ocel::{
+use crate::core::{
+    event_data::object_centric::{
         linked_ocel::{IndexLinkedOCEL, LinkedOCELAccess, SlimLinkedOCEL},
-        ocel_struct::{OCELEvent, OCELRelationship, OCELType},
+        OCELEvent, OCELRelationship, OCELType,
     },
     OCEL,
 };
