+++ process_mining/src/discovery/mod.rs
@@ -3,3 +3,4 @@
 //! Process discovery algorithms typically learn process model representations
 //! from input event data.
 pub mod case_centric;
+pub mod object_centric;
