+++ process_mining/src/discovery/object_centric/mod.rs
@@ -1,0 +1,2 @@
+//! Object-centric Process Discovery Techniques
+pub mod oc_declare;
